% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-modGenCompare.R
\name{modGenCompare}
\alias{modGenCompare}
\title{compare two models}
\usage{
modGenCompare(modelRun1, modelRun2)
}
\arguments{
\item{modelRun2}{example:
modGenCompare(modelRunLinear, modelRunBayes)}
}
\description{
compare two models
}

