% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-modGenCompareAvm.R
\name{modGenCompareAvm}
\alias{modGenCompareAvm}
\title{compare two avms}
\usage{
modGenCompareAvm(modelRun1, modelRun2, dateVar1, dateVar2, dy = FALSE)
}
\arguments{
\item{dy}{= T or F
version: 0.3
example:
modGenCompareAvm(modelRunLinear, modelRunBayes, "period", "period", dy=TRUE)
modGenCompareAvm(modelRunLinear, modelRunBayes, "period", "period")}
}
\description{
compare two avms
}

