% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-Fns.R
\name{modGenCreate}
\alias{modGenCreate}
\title{create model object: creates model object}
\usage{
modGenCreate(modelType = "", modelSet = data.frame(),
  modelParams = list(), modelVars = list())
}
\value{
list
version 0.3
example: modGenCreate(modeType=modeType, modelSet=modelSet, modelParams=modelParams, modelVars=modelVars)
}
\description{
create model object: creates model object
}

