% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-modGenExportContrib.R
\name{modGenExportContrib}
\alias{modGenExportContrib}
\title{export contributions in standard format}
\usage{
modGenExportContrib(modelRun, dateVar, filename = "./contribTab.csv")
}
\description{
export contributions in standard format
}

