% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init-Fns.R
\name{modGenRun}
\alias{modGenRun}
\title{run model object: creates output run on model object}
\usage{
modGenRun(modelObject)
}
\value{
list
example: 
modGenRun(modelObject=modelObject)
version 0.3
}
\description{
run model object: creates output run on model object
}

