% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-modGenPlotAvm.R
\name{modGenPlotAvm}
\alias{modGenPlotAvm}
\title{create model avm}
\usage{
modGenPlotAvm(modelRun, dateVar, dy = FALSE)
}
\arguments{
\item{dy}{version: 0.3
example:
modGenPlotAvm(modelRunLinear, "period")
modGenPlotAvm(modelRunBayes, "period")}
}
\description{
create model avm
}

